requirements.txt
wheel
Django
djangorestframework
paho-mqtt
psutil
mysqlclient
gunicorn
influxdb-client
django-celery-beat
pandas